requirements.txt
asgiref==3.8.1
Django==5.1.1
drf-yasg==1.21.7
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
